Compare same datatypes only #37 This breaks on compiling on El Capitan

+++ macosvpn/Classes/VPNArguments.h
@@ -26,6 +26,6 @@
 
 + (NSArray<NSString *>*) serviceNames;
 + (NSArray<VPNServiceConfig*>*) serviceConfigs;
-+ (NSUInteger) command;
++ (UInt8) command;
 
 @end
